Skip ES compatibility checks for SNAPSHOT builds
